requirements.txt
amqp==1.4.5
anyjson==0.3.3
astroid==1.1.1
billiard==3.3.0.17
celery==3.1.11
click==1.1
jsonpickle==0.7.1
kombu==3.0.16
logilab-common==0.61.0
mercurial==3.0
paho-mqtt==0.9
pep8==1.4.6
psycopg2==2.5.3
pylint==1.2.1
pymongo==2.7
pytz==2014.3
pyzmq==14.3.0
stevedore==0.14.1
wsgiref==0.1.2
